+++ Tests/Runtime/XRSimulation/SimulationCameraTests.cs
@@ -31,8 +31,8 @@ namespace UnityEngine.XR.Simulation.Tests
             var xrCamera = xrOrigin.Camera;
             Assert.IsNotNull(xrCamera);
 
-            var poseProvider = FindObjectsUtility.FindAnyObjectByType<SimulationCamera>();
-            Assert.IsNotNull(poseProvider, $"No active {nameof(SimulationCamera)} is available.");
+            var poseProvider = FindObjectsUtility.FindAnyObjectByType<SimulationCameraPoseProvider>();
+            Assert.IsNotNull(poseProvider, $"No active {nameof(SimulationCameraPoseProvider)} is available.");
         }
     }
 }