+++ CHANGELOG.md
@@ -8,6 +8,13 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [5.2.0] - 2025-03-24
+
+### Added
+
+- Added enum `ARDebugMenu.DebugMenuRenderMode` and property `ARDebugMenu.debugMenuMode` that allows you to specify the render mode of the menu's canvas.
+- Added XR Simulation's `SimulationCameraPoseProvider` as a public component.
+
 ## [5.1.6] - 2024-12-11
 
 ### Fixed